A
.git/COMMIT_EDITMSG[+] [unix] (02:03 11/12/2025)                         1,2 Top
-- INSERT --

Merge remote-tracking branch 'upstream/master'

modified:   src/main/java/ui/ft/ccit/faculty/transaksi/barang/BarangService.java

+++ src/main/java/ui/ft/ccit/faculty/transaksi/barang/Barang.java
@@ -14,10 +14,10 @@ public class Barang {
     private String nama;
 
     @Column(name = "stok")
-    private Integer stok;
+    private Short stok;
 
     @Column(name = "harga")
-    private Integer harga;
+    private Double harga;
 
     @Column(name = "persen_laba")
     private Double persenLaba;
@@ -35,7 +35,7 @@ public class Barang {
         // for JPA
     }
 
-    public Barang(String idBarang, String nama, Integer stok, Integer harga,
+    public Barang(String idBarang, String nama, Short stok, Double harga,
             Double persenLaba, Double diskon,
             Byte idJenisBarang, String idPemasok) {
         this.idBarang = idBarang;
@@ -66,19 +66,19 @@ public class Barang {
         this.nama = nama;
     }
 
-    public Integer getStok() {
+    public Short getStok() {
         return stok;
     }
 
-    public void setStok(Integer stok) {
+    public void setStok(Short stok) {
         this.stok = stok;
     }
 
-    public Integer getHarga() {
+    public Double getHarga() {
         return harga;
     }
 
-    public void setHarga(Integer harga) {
+    public void setHarga(Double harga) {
         this.harga = harga;
     }
